requirements.txt
Django==4.2.11
djangorestframework==3.14.0
gunicorn==21.2.0
mysqlclient==2.2.1